bedrock-intake/requirements.txt
boto3>=1.34.0
PyGithub>=2.1.1
hypothesis>=6.92.0
pytest>=7.4.3
pytest-cov>=4.1.0
